Fix freeze when stopping CPU recording on a pre-O device
When multiple devices are connected, ddmlib's static method
ClientData#setMethodProfilingHandler will be called multiple times.
This results in interference with pre-O CPU recording from the
legacy pipeline's CPU service proxy.

The fix skips profiler service proxy registration if we're on the
unified pipeline. For CPU and memory proxies, their functionalities
are already implemented as commands. The rest of the proxies are
just pass-throughs.

Fixes: 177612469
Test: updated
Change-Id: I11e402b2160b9cc476f6dde4946625a5d2ba901f