Remove use of JUnit3RunnerWithInners
JUnit3RunnerWithInners was created to make it easier to run the tests
from all inner classes of an outer class, but (1) it does not seem
to work for us in the IDE, and (2) it does not affect which tests
end up running in Bazel.

So, it is safe to remove. Removing it helps reduce our dependency
on the Kotlin plugin test framework.

Bug: 178713991
Test: existing android-kotlin tests
Change-Id: I0948629ee7cdeeda401ece4280d6a7bc99cf9d95