[Dsl,.gr] Don't escape explicit hierarchical names of blocks
This problem only currently manifests in Groovy because of the
difference in the association of elements in the Psi hierarchy, where
Kotlin has a right-leaning tree for dotted properties while Groovy has a
left-leaning tree.  In the case where we were giving an already-extant
block (such as android->defaultConfig) a nameElement (because the
previous uses had been in statements only), we were escaping the
hierarchical dots, if any were present.

We have the apparatus to rename hierarchical names correctly, so use it.

Fixes: 178801999
Test: added
Change-Id: I1f36ad3a80752ca6700becc194a9ef9cdc95556b