[Dsl] Support parsing augmented assignment
Define a new AUGMENTED_ASSIGNMENT ExternalNameSyntax, and decorate the
resulting element with that as its syntax when parsing a
KtBinaryExpression (Kotlin) / GrAssignmentExpression (Groovy) of the
appropriate form.

Adjust the name canonization performed by blocks for elements they know
about to handle VAL properties of type MUTABLE_SET.

Define and use a new augmentParsedExpression method which implements
augmentation semantics, and perform canonization in the
GradleDslBlockElement implementation of it.

Bug: 177902134
Bug: 179483658
Test: added to tests of packagingOptions sub-blocks
Change-Id: Iaace9489bc819114a05e8d7aa1b4728661becfe1