[Dsl] Insert new Ext block after Plugins
Gradle requires that there be essentially no Dsl textually preceding
plugins declarations, so that the plugins can be loaded to perform the
rest of the build configuration.  Ensure that any new Ext block is after
a Plugins block, if any, as well as after an Apply.

This change might tickle the reviewer's danger sense.  Why do we not
need to worry about being after *both* of the Apply and Plugins blocks?
Well, possibly we do, but this should be relatively infrequent:
programmatic addition of plugins will use whichever is there first, so
a project with combined Apply and Plugins would have to have both
explicitly, which would be weird.

However, there are more ways that this can go wrong.  Our implementation
of allprojects(), and its close cousin subprojects(), involves
pre-populating the properties of a new GradleBuildFile with the
properties of the parent's subprojects().  There is logic in
GradlePropertiesDslElement.ElementList to convert from an intended
index (in this case 0 or 1) to a "real" index in the properties list,
skipping over any that are APPLIED (or TO_BE_REMOVED); however, the
population of the properties subprojects() places those properties in
EXISTING state, which is probably wrong.  Unfortunately, it's not quite
straightforward to change the population to use APPLIED state (which is
closest to what is actually going on) because that bypasses some logic
to handle top-level language level properties.  All of this to explain
why this change does not claim to fix the associated bug.

Bug: 174827009
Test: added to PropertyOrderTest
Change-Id: I2a304c38249f59f6794347758736185f5edba170