Fix editor losing focus when Animation Preview is enabled
After recently changing the animation preview to use an IntelliJ tabs
component, we started to request the focus on the timeline when adding a new
tab to make sure the tab containing the timeline had the focus. There is a
couple of issues with that:

1) It's not enough. If you change the focus to the editor after adding a tab
and then click the timeline again, the tab will not get the focus.
2) When live literals is enabled, every change in a literal will cause the
animation preview to update and create new tabs corresponding to the
up-to-date animations, changing the focus from the editor to the animation
preview. This can be annoying if the user is trying to edit multiple
literals before getting their attention to the animation preview.

Both concerns are fixed by this CL, which simply requests the focus to the
tab containing the timeline when the user clicks on it.

Fixes: 180700119
Test: Manually verified by editing animation with live literals enabled
Change-Id: I31738efff890a0dfdee670c5f6ea46843af61d3f