Show weak warnings for transition properties label param
Similar to what we do with updateTransition calls, we need to show weak
warnings if the user does not set the label field when defining transition
properties. If they don't explicitly set the label, a default value will be
used to represent the property, and that might make it harder to inspect it
in the animation preview. For example, it would be better for the user to
see "Corner Radius" and "Scale" instead of "FloatAnimation" twice.

Fixes: 181986045
Test: AnimationInspectionsTest
Change-Id: Ia329401e079544121e3d2b996114c382eef2ad2e