Stop launching new clients when inspector is minimized
Also, fixed an issue where the initial state change of the window was
ignored if the window started out open. This is because the listener
always treated the initial state an hidden, so when the user went to
minimize the window for the first time, it thought the state went
from hidden -> hidden, and ingored it.

Fixes: 180958559
Test: Added
Change-Id: I597d6d25195e36b6ae4be3180cfe8c1ba1dc171c