Revert "Let developers background the screen recorder task"
This reverts commit faef11a843a81cf8cda8357ca8bffbe207cefd8c.

Between Studio 4.1 and 4.2, backgroundable progress dialogs broke in the
IntelliJ platform. Users aren't able to interact with the parent Studio
window after they background the progress dialogs.

When I merge this commit users WILL NOT be able to interact with the
Emulator tool window when the virtual device's screen is being recorded.
This is intentional and the team will revisit that issue.

Bug: 180474805
Test: Not applicable
Change-Id: I037c19d4db9bc0138103eef040d1c35464d77e40