[AUA] Initialize the Java8Default action to Accept
For non-interactive use, the safe default is to insert defaults to
preserve existing behaviour, because that is more likely to be
the (locally) correct decision.  For interactive use, we default to the
opposite, presuming that if something goes wrong the user can undo, but
that this possible inconvenience is outweighed by the benefits of
getting more projects to use Java 8.  (This matches the existing
Dialog-based behaviour)

Bug: 178569506
Test: existing
Change-Id: I40199f77104465186e80cd5f51f0fee98e92c7d6