Add independent tree nodes to component tree
This change is a preparation for performing the system view filter
in Studio rather than making another tree request to the agent.

Add a separate tree of tree nodes to each ViewNode, and build the
tree structure after each structural change in the model.
Perform the filtering of the DecorView node by eliminating that
node from the tree nodes if system views are hidden.

Bug: 182838650
Test: Existing tests updated
Change-Id: Idb9de6dfe3253042a534dcc0b2927c35079d11e1