[AUA] Add some smoke tests for Tool Window model
In order to start making the Tool Window model testable, we make some
minor changes to visibility, and also change the use of threads when in
unit testing mode.  This does not alter the behaviour of the system in
production, but allows inclusion of some basic tests of the model after
the tool window is initialized.

Bug: 183107211
Test: added in this CL
Change-Id: Ic8aa892bf181ca7542003afba7517f65f7d3a3bb