Fix bounds navigation
This adds a workaround for a bug in the Compose Runtime where we will
not be sent a packageHash for a valid file.
We will now only use the packageHash to disambiguate if needed. If there
is only one file, we will navigate to that one.

Test: Reenables tests for navigation
Bug: N/A
Change-Id: I97a101d4c9b0c2f00e36cd98d951c666b6c80318
(cherry picked from commit 99224b99f422198ed935ecdfca623c03e8856268)