[AUA] Support Plugins Dsl when upgrading other plugins
When we upgrade plugins because of a known incompatibility with AGP or
Gradle, check and handle cases where the plugins are declared in a
plugins { ... } block in a build file using the Dsl, as opposed to
classpath dependency configurations in the buildscript block.

This does not yet handle the declaration of plugins in the settings
file, for which we need a little bit of additional support from the Dsl
layer.

Bug: 185115827
Test: added
Change-Id: I2b0873058151ca805a55824a522ec30af068a870