Use HelpTooltip for tooltip in device dropdown.
We want to have a consistent UI for tooltips. For that we need to use HelpTooltip.
We need tooltip for button in DeviceAndSnapshotComboBoxAction and for cells in a list of DeviceAndSnapshotComboBoxAction.

Intellij uses HelpTooltip for all component tooltips through HelpTooltipManager
or you can manually install HelpTooltip on a component/action.

First approach (delegate tooltip to IDEA and allow it to use HelpTooltipManager) doesn't work because
we need a title and a description in tooltip (HelpTooltip().setTitle(..).setDescription(..)), but HelpTooltipManager
installs only title. See HelpTooltipManager.showTooltip.

The second approach doesn't work for a button in DeviceAndSnapshotComboBoxAction because HelpTooltip has fixed text for component. In our case tooltip text for a button in DeviceAndSnapshotComboBoxAction is dynamic and depends on chosen device. And a second approach doesn't work for a list because we can install in only for the whole list, not a cell.

In this CL I introduce UpdatableDeviceHelpTooltip that I use for a button and a list in DeviceAndSnapshotComboBoxAction.
We can update UpdatableDeviceHelpTooltip depending on which device is chosen (for button) or which device is hovered for list.

The old mechanism for tooltip via presentation.putClientProperty(JComponent.TOOL_TIP_TEXT_KEY, ..) is gone.

Fixes: 184914615
Test: manually
Change-Id: I13a1c5150f9850c25a18237fde49d60337dafb45