Reduce ExternalLibrary to ExternalAndroidLibrary
which is used to describe Android library resources.

This interface does not describe JARs contained in a library in a way
that can describe Gradle dependencies it is unlikely that we can easily
have a universal description that works for all project systems. Instead
we have higher level more specific APIs which make more sense for
different project systems.

This change renames `ExternalLibrary` to `ExternalAndroidLibrary` as
it is never used to access details of simple `.jar` libraries. Moreover,
it is only used to find details about Android assets and resources
contained in a library.

This change also results in creating a smaller number of objects of
a smaller size.

Bug: n/a
Test: existing updated
Change-Id: I5861e2ef376472bece4a7b1b879b139492920b07