Remove warning log spam from LayoutlibSceneManager
In LayoutlibSceneManager#executeCallbacksAndRequestRender(), we set a
timeout to the action executed in the render thread. If the render task
can't be executed on time, the function logs the exception and returns
false. Using warning is not really needed here, since the caller will check
for the function return value. Furthermore, since this function might be
called too often (e.g. Animation Preview calls it once per clock time
change, which is continuous while the animation is playing), we might end up
spamming the idea logs.

Bug: N/A
Test: N/A
Change-Id: I36b8c946b0f352c88e07e2d19b3a57872bf0468f