[AUA] Remove an error log call
The situation is not an error, and in fact can happen fairly easily (and
might even happen systematically on initialization).  If two refreshes
happen fairly quickly, two calls to refreshTree() will be enqueued but
it is possible for neither to run until both of the refresh() calls have
cleared the tree and returned, so the second refreshTree() will see a
non-empty tree model.

Bug: n/a
Test: manual
Change-Id: Ie82827df07060f82809bb13a860c6b01c800e5a3
(cherry picked from commit 370561da161ad55b8f25313e2594349753a74188)