Format compose refresh elapsed time in Event Log
We were using an experimental kotlin API to properly format the Compose
refresh elapsed time, which we write to Android Studio's Event Log. This API
has changed and broke compatibility with Kotlin plugin 1.5.0. This issue was
addressed by change Ic79d395f313ef, but there we hardcoded the log to always
use milliseconds, which is not ideal for longer refreshes, which can take
several seconds.

This CL adds formatting back, this time using DurationFormatUtils, which has
a stable API.

Bug: N/A
Test: Manually checked the event log
Change-Id: Ieb8beb72737ebd8160b711fe43954dd5f46a43ef