Fix resolving wrong ViewStub class type in view binding
There is code to check what class a <ViewStub> xml element should expand
to, which is "android.view.ViewStub" most of the time but something like
"android.databinding.ViewStubProxy" in data binding.

The current code is a bit too general - it always returns the data
binding version if data binding is enabled for the current module.
However, a project can have view binding AND data binding enabled for
the same module, so we need to take the current XML context into account
to decide what the final class resolution should be.

Fixes: 186306088
Test: Covered by existing + Updated
Change-Id: I17169819966a9b1ca658572f602a65c2359be9f3