Fix animation label quick-fix when only lambda param is set
We were relying on the valueArgumentList property of transition animation
animate* APIs when trying to apply a quick fix to the weak warning shown
when this APIs are called without the label parameter set. The problem is
the other argument of these APIs is usually a lambda, which is not returned
as part of valueArgumentList, but part of lambdaArguments instead. Calling
getOrCreateValueArgumentList() instead of valueArgumentList directly fixes
this issue.

Fixes: 186794675
Test: Added regression test to AnimationInspectionsTest
Change-Id: I20ff36d04223acd518c92692ce66d56152f3b672