[Dsl] Collect Dsl information into a custom object
Until now we have been using an ImmutableMap with SurfaceSyntax keys and
ModelEffectDescription values.  This has worked because there has been
no need to consider a many-to-many mapping from surface syntax to
model -- any given surface syntax has been treated as corresponding to
at most one model effect.  (The converse has never been the case: there
have always been multiple SET methods for a given model property,
e.g. setFoo() and foo() for a Dsl property `foo`.)

However: we barely use the map as a map, getting a value for a surface
syntax in only one place -- and even there, there is a realisation that
the map abstraction is not quite right, because although a lookup for
property syntax works, the same is not the case for method syntax,
because of variadic functions (an expressed arity can match more than
one encoded arity).  And with the advent of version-awareness in the
Dsl, there will be full many-to-many relations, as surface syntax can
mean different things in different versions of the Dsl.

Instead, collect into a custom ExternalToModelMap object.  At the moment
this is a bare wrapper around an (ordered) Set, each element of the set
corresponding to one row of information in the Dsl table.  We will in
the near future be able to add a version constraint expressing the
validity of each row in any given verison of the Dsl.

Bug: 186623625
Test: existing (no change in functionality)
Change-Id: I563b7aa39f3569944507bd5815ac5da5c2949d3a