Reverts Id53294b2aaff1d300b07fe02a0a6a201eb493cc5
When "run instrumentation tests using Gradle" option
is enabled, we can skip running make-before-run-task
as an optimization by not building the same app
and test APKs twice. In practice, the build doesn't
occur twice thanks to the build cache although there
could be a case where the build cache doesn't work.

The "run instrumentation tests using Gradle" option
is still in opt-in feature and we expect a user
moves back-and-forth to try the new feature and we
don't want to mess up with their generated run
configuration due to this settings.

Fixes: 186582479
Test: n/a
Change-Id: I99456a6f460e5a78e1bd99b50fc18b5e73f5c6b7