Uniquely name downloaded inspector jars
We used to keep the inspector.jar name from all bundled inspectors,
however this created a problem when two or more such inspectors are
pushed to the device, in which case one would overwrite the other.

This change uniquely names the inspector jar once it is resolved. The
name is of the format (group_id)-(artifact_id)-(version).jar

Test: updated existing test.
Fixes: 188053833
Change-Id: I09e335889ea88be60d16281d40c57a49fea4eafd