[BA] Make details copyable
Make possible to select details text defined in html components
and copy it.
The problem was that this requires editor pane to be focusable
but it does not report focus properly resulting in bad UX.
The solution applied in this CL is to define focus traversal policy
that would not accept JEditorPane as focused component. Since we use
it only for html text rendering in BA UI it is quite enough. Once we
need to enhance it somehow we can adjust the filter criteria as needed.

Bug: 156912156
Test: N/A
Change-Id: I36eae0acf26da772919e08a5a6948fd1336905eb