Adjust LambdaResolver for lambda with multi line call
When the last expression in a lambda is a multi line call, the line
range we get from JVMTI does not include the end paranthesis of the
call nor the end bracket of the lambda.
Adjust the algorithm for finding the best matching lambda.

This was the case in b/184905448 where the Compose implementation of
a LazyList calls SubcomposeLayout with a lambda that contains a call
as the return value of the lambda.

Fixes: 184905448
Test: Added unit test
Change-Id: If0108297d3db21b2a8fd29574fbf486f821b0166