Include some transitive dependencies for manifest merger
To build the final merged manifest the manifest merging requires to know
all app dependencies including any transitively included libraries which
are not on the compile class path.

This change adds any libraries transitively included from depency
modules directly to the list of manifest merger sources. This change,
however, does not include any transitive dependencies which are not on
the compile class path of any modules.

Bug: 178796185
Test: n/a
Change-Id: I247f87fd76672e1b8b24179f28cf405b306ad2ce