[AUA] Fix another test
On the Mac builder, there currently is apparently no version of AGP
available later than 4.2.0.  This makes the test checking that a
third-party plugin is marked for upgrade (and a separate entry in the
tree) fail, because in an upgrade to 4.2 it does not need that.  Make
the version of the plugin used in the test even earlier, such that even
upgrades to 4.2.0 require an upgrade to the plugin version.

Fixes: 189890821
Test: this is it
Change-Id: I00cd4bdbdb16ade1197e126ae9dc7d59cc12f010