Database Inspector: allow limiting response size
Added a support to limit query response size (in bytes).

This allows us to:
- avoid OutOfMemoryError for queries returning a lot of data
- minimize overhead of back/forth communication by maximising
  the response size

Cherry-pick justification:

This fixes an issue found in stress testing - in scenarios where the
response would be large (e.g. 200 records of 0.5 MB each) we were
observing OutOfMemoryErrors (OOME). These are not currently surfaced to
the user (result of using vanilla Futures).

An alternative to the fix (hardcoding row count limit in response) not
viable as would have to be set low (e.g. 10 rows, so that we can handle
10 x 2MB = 20MB) which would make the tool unusable for e.g. 100,000
small records.

Bug: 161081452
Test: ./gradlew :sqlite:sqlite-inspection:cC

Change-Id: I2a2144680c8aff6de5b7b6506d35f567fa4ac450